requirements-unit.txt
alembic==1.5.4
appdirs==1.4.4
attrs==20.3.0
bcrypt==3.1.6
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
distlib==0.3.1
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.7.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
igramscraper @ git+https://github.com/realsirjoe/instagram-scraper.git@ae7824b6abab6971d3749a3316471ad77a1c7d07
importlib-metadata==2.1.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.4
more-itertools==8.6.0
Naked==0.1.31
numpy==1.19.4
packaging==20.9
pandas==1.1.5
parse==1.18.0
Pillow==8.1.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pyee==7.0.4
PyJWT==1.4.2
PyMySQL==1.0.2
pyparsing==2.4.7
pyppeteer==0.2.3
pyquery==1.4.3
pytest==6.2.2
pytest-filter-subpackage==0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-slugify==3.0.2
pytrends==4.7.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-html==0.10.0
selenium==3.141.0
selenium-stealth==1.0.6
shellescape==3.8.1
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.22
text-unidecode==1.2
toml==0.10.2
tqdm==4.55.1
typing-extensions==3.7.4.3
urllib3==1.26.3
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==0.16.1
WTForms==2.2.1
zipp==3.4.1
